% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{assignDomainSyntax}
\alias{assignDomainSyntax}
\title{Assign Domain Syntax}
\usage{
assignDomainSyntax(object)
}
\arguments{
\item{object}{An object that will store the domainTable and domainStartDate names, requires a domainId attribute assigned}
}
\value{
An object with domainTable and domainStartDate assigned
}
\details{
Assigns the domainTable and start date syntax based on a domainId
}
\author{
Ajit Londhe
}
