% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{buildCovariate}
\alias{buildCovariate}
\title{Build Covariate}
\usage{
buildCovariate(label, covariateTable = NULL, scoreColumn = NULL,
  categoryColumn = NULL, conceptIds = NULL, domainId = NULL)
}
\arguments{
\item{label}{The name of the covariate}

\item{covariateTable}{The name of the table holding covariate scores or categories}

\item{scoreColumn}{(Optional) The name of the column that holds the covariate's continuous score (if available)}

\item{categoryColumn}{(Optional) The name of the column that holds the covariate's categorical levels (if available)}

\item{domainId}{(Optional) The domain Id (as defined in the OMOP Vocabulary) of the concept}

\item{conceptId}{(Optional) The Concept Ids of the covariate}
}
\value{
A covariate object
}
\details{
Builds a Covariate for supporting evidence
}
\author{
Ajit Londhe, Jamie Weaver
}
