% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{buildDatabase}
\alias{buildDatabase}
\title{Build Database}
\usage{
buildDatabase(id, dbms, dbName, server, port, cdmDatabaseSchema = NULL,
  resultsDatabaseSchema = NULL, scratchDatabaseSchema = NULL, user,
  password, connectionDetails = NULL)
}
\arguments{
\item{id}{The unique key for the database}

\item{dbms}{The database platform type ("sql server," "pdw," "redshift," "oracle," "mysql")}

\item{dbName}{The technical name of the database}

\item{server}{The name of the server host}

\item{port}{The SQL port of the server host}

\item{cdmDatabaseSchema}{The fully qualified schema name of the CDM schema}

\item{resultsDatabaseSchema}{The fully qualified schema name of the results schema}

\item{scratchDatabaseSchema}{The fully qualified schema name of the scratch schema}

\item{user}{The user name of the SQL account}

\item{password}{The password of the SQL account}

\item{connectionDetails}{The connectionDetails object for this database}
}
\value{
A database object
}
\details{
Builds a database object for use in the study
}
\author{
Ajit Londhe
}
