% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createBalanceFile}
\alias{createBalanceFile}
\title{Create Balance File}
\usage{
createBalanceFile(run_id, cdmDb, dcoList, cmOutputPath, reportsOutputPath,
  onlyUnbalanced = FALSE)
}
\arguments{
\item{run_id}{The run id of this execution}

\item{cdmDb}{The CDM database object constructed with \code{buildDatabase}}

\item{dcoList}{The list of drug-comparator-outcomes}

\item{cmOutputPath}{The path to the CohortMethod files}

\item{reportsOutputPath}{The path to the reports file root}

\item{onlyUnbalanced}{Should only covariates with SMD > 0.1 be included?}
}
\details{
Creates covariate balance file for a CDM's analyses
}
\author{
Ajit Londhe, Jamie Weaver
}
