% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createModelSummary}
\alias{createModelSummary}
\title{Create Model Summary}
\usage{
createModelSummary(cdmDb, dcoList, cmOutputPath, negativeControlConceptIds,
  reportsOutputPath, subGroupList = NULL, includeNegControls = FALSE)
}
\arguments{
\item{cdmDb}{The CDM database object constructed with \code{buildDatabase}}

\item{dcoList}{The list of drug-comparator-outcomes}

\item{cmOutputPath}{The path to the CohortMethod files}

\item{negativeControlConceptIds}{List of negative control concept Ids}

\item{reportsOutputPath}{The path to the reports file root}

\item{subGroupList}{The list of subgroups, built with \code{buildSubGroup}}

\item{includeNegControls}{Should the negative control outcome models be exported?}
}
\details{
Creates the model summary files for a specific CDM
}
\author{
Ajit Londhe, Jamie Weaver
}
