% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createOutcomeModelPlots}
\alias{createOutcomeModelPlots}
\title{Create outcome model visualizations}
\usage{
createOutcomeModelPlots(outcomeIds, reportsOutputPath, cmOutputPath)
}
\arguments{
\item{outcomeIds}{An array of the outcome ids}

\item{reportsOutputPath}{The path pointing to the folder that should hold charts and tables}

\item{cmOutputPath}{The path pointing to the folder that holds the CohortMethod assets}
}
\details{
Creates follow-up time plots, KM curves, RR forest plots
}
\author{
Jamie Weaver, Ajit Londhe
}
