% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{createTable1}
\alias{createTable1}
\title{Get Table 1}
\usage{
createTable1(webApiPrefix, webApiUseSsl = FALSE, cdmDb, scratchDb,
  scratchTablePrefix, dcoList, cmOutputPath)
}
\arguments{
\item{webApiPrefix}{The URL prefix of the WebAPI instance}

\item{webApiUseSsl}{Does the WebAPI use HTTPS?}

\item{cdmDb}{The CDM database object constructed with \code{buildDatabase}}

\item{scratchDb}{The scratch database object constructed with \code{buildDatabase}}

\item{scratchTablePrefix}{The prefix of all scratch tables created}

\item{dcoList}{The list of drug-comparator-outcomes}

\item{cmOutputPath}{The path to the CohortMethod files}
}
\details{
Generates and exports a Table 1 for all drug-comparator-outcomes
}
\author{
Ajit Londhe, Jamie Weaver
}
