% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmExecute.R
\name{executeFullCM}
\alias{executeFullCM}
\title{Execute Full CohortMethod}
\usage{
executeFullCM(connectionDetails, cdmDatabaseSchema, dcoList,
  excludedCovariateConceptIds, cmOutputPath, exposureDatabaseSchema,
  exposureTable, outcomeDatabaseSchema, outcomeTable, fitOutcomeModel,
  excludeCovariateIds, deleteCovariatesSmallCount,
  errorOnHighCorrelation = TRUE, stopOnError = TRUE)
}
\arguments{
\item{connectionDetails}{The connectionDetails object}

\item{cdmDatabaseSchema}{The fully qualified name of the CDM database schema}

\item{dcoList}{The list of drug-comparator-outcomes}

\item{excludedCovariateConceptIds}{The list of covariate concept Ids to exclude}

\item{cmOutputPath}{The path to the CohortMethod files}

\item{exposureDatabaseSchema}{The fully qualified schema name for the exposures}

\item{outcomeTable}{The name of the table with the outcome cohorts}

\item{fitOutcomeModel}{Should the outcome model be fit?}

\item{errorOnHighCorrelation}{Should CohortMethod throw an error if highly correlated covariates are detected?}

\item{stopOnError}{Should CohortMethod stop on an error?}
}
\details{
Executes Cohort Method using MultipleAnalyses
}
\author{
Ajit Londhe, Jamie Weaver
}
