% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{generateAgeGroups}
\alias{generateAgeGroups}
\title{Generates Age Groups}
\usage{
generateAgeGroups(dcoList, cdmDb, scratchDb, scratchTablePrefix)
}
\arguments{
\item{dcoList}{The list of drug-comparator-outcomes}

\item{cdmDb}{The CDM database object constructed with \code{buildDatabase}}

\item{scratchDb}{The scratch database object constructed with \code{buildDatabase}}

\item{scratchTablePrefix}{The prefix of all scratch tables created}
}
\details{
Generates age groups
}
\author{
Ajit Londhe, Jamie Weaver
}
