% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generateNegativeControlCohorts}
\alias{generateNegativeControlCohorts}
\title{Generate Negative Controls}
\usage{
generateNegativeControlCohorts(cdmDb, scratchDb, dcoList, scratchTablePrefix,
  negativeControlConceptIds, studyStartDate)
}
\arguments{
\item{cdmDb}{The CDM database object constructed with \code{buildDatabase}}

\item{scratchDb}{The scratch database object constructed with \code{buildDatabase}}

\item{dcoList}{The list of drug-comparator-outcomes}

\item{scratchTablePrefix}{The prefix of all scratch tables created}

\item{negativeControlConceptIds}{The negative control concept Ids}

\item{studyStartDate}{The study start date}
}
\details{
Generates the outcome cohorts needed for negative controls
}
\author{
Ajit Londhe, Jamie Weaver
}
