% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generatePsOutputs}
\alias{generatePsOutputs}
\title{Create PS Outputs}
\usage{
generatePsOutputs(connectionDetails, dcoList, cdmDb, scratchDb,
  scratchTablePrefix, cmOutputPath, reportsOutputPath, webApiPrefix,
  webApiUseSsl, numTopCovariates = 20, imageExtension = ".png",
  insertCohortsToDb = FALSE)
}
\arguments{
\item{connectionDetails}{The connectionDetails object}

\item{dcoList}{The list of drug-comparator-outcomes}

\item{cdmDb}{The CDM database object constructed with \code{buildDatabase}}

\item{scratchDb}{The scratch database object constructed with \code{buildDatabase}}

\item{scratchTablePrefix}{The prefix of all scratch tables created}

\item{cmOutputPath}{The path to the CohortMethod files}

\item{reportsOutputPath}{The path to the reports file root}

\item{webApiPrefix}{The URL prefix of the WebAPI instance}

\item{webApiUseSsl}{Does the WebAPI use SSL?}

\item{numTopCovariates}{How many covariates should be included in the Top Covariates balance graph?}

\item{imageExtension}{What file image extension to use?}

\item{insertCohortsToDb}{Should the PS matched cohorts be inserted to the scratch database?}
}
\details{
Creates balance plots, writes propensity-matched data to scratch
}
\author{
Ajit Londhe
}
