% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getCohortDefName}
\alias{getCohortDefName}
\title{Get Cohort Definition Name}
\usage{
getCohortDefName(webApiPrefix, cohort_definition_id,
  webApiwebApiUseSsl = FALSE)
}
\arguments{
\item{webApiPrefix}{The URL for WebAPI}

\item{conceptSetId}{The concept set id in Atlas}

\item{webApiUseSsl}{Does the WebAPI instance use SSL?}
}
\value{
A list of concept Ids
}
\details{
Obtains the name of a cohort
}
\author{
Ajit Londhe
}
