% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getFormattedCohortName}
\alias{getFormattedCohortName}
\title{Get Formatted Cohort Name}
\usage{
getFormattedCohortName(webApiPrefix, treatmentId, webApiUseSsl = FALSE)
}
\arguments{
\item{webApiPrefix}{The URL prefix of the WebAPI instance}

\item{treatmentId}{The cohort definition Id}

\item{webApiUseSsl}{Does the WebAPI instance use SSL?}
}
\value{
A string with the formatted cohort name
}
\details{
Gets a formatted name for a cohort
}
\author{
Ajit Londhe
}
