% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getIncludedConcepts}
\alias{getIncludedConcepts}
\title{Get Included Concepts}
\usage{
getIncludedConcepts(webApiPrefix, conceptSetId, webApiUseSsl = FALSE)
}
\arguments{
\item{webApiPrefix}{The URL for WebAPI}

\item{conceptSetId}{The concept set id in Atlas}

\item{webApiUseSsl}{Does the WebAPI instance use SSL?}
}
\value{
A list of concept Ids
}
\details{
Obtains the full list of concept Ids in a concept set
}
\author{
Ajit Londhe
}
