% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{initializeEnvironment}
\alias{initializeEnvironment}
\title{Initialize Environment}
\usage{
initializeEnvironment(scratchTablePrefix)
}
\arguments{
\item{scratchTablePrefix}{A prefix to help tag the FF files}
}
\details{
Initializes the study environment
}
\author{
Ajit Londhe
}
