% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runNegativeControls}
\alias{runNegativeControls}
\title{Run Negative Controls}
\usage{
runNegativeControls(cdmDb, dcoList, scratchDb, connectionDetails,
  scratchTablePrefix, negativeControlConceptIds, excludedCovariateConceptIds,
  cmOutputPath, webApiPrefix, deleteCovariatesSmallCount,
  fitOutcomeModel = TRUE, cdmVersion = 5)
}
\arguments{
\item{cdmDb}{The CDM database object constructed with \code{buildDatabase}}

\item{dcoList}{The list of drug-comparator-outcomes}

\item{scratchDb}{The scratch database object constructed with \code{buildDatabase}}

\item{connectionDetails}{The connectionDetails object}

\item{scratchTablePrefix}{The prefix of all scratch tables created}

\item{negativeControlConceptIds}{The negative control concept Ids}

\item{excludedCovariateConceptIds}{The concept Ids to exclude in Feature Extraction}

\item{cmOutputPath}{The path to the CohortMethod files}

\item{webApiPrefix}{The URL prefix of the WebAPI instance}

\item{deleteCovariatesSmallCount}{The minimum cell size}

\item{fitOutcomeModel}{Should the outcome models be fit?}

\item{cdmVersion}{The CDM version of the CDM database}
}
\details{
Executes CohortMethod for negative control outcomes 
                                   by invoking \code{executeFullCm}
}
\author{
Ajit Londhe, Jamie Weaver
}
