% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runPrimaryAnalyses}
\alias{runPrimaryAnalyses}
\title{Run Primary Analyses}
\usage{
runPrimaryAnalyses(cdmDb, dcoList, connectionDetails,
  excludedCovariateConceptIds, cmOutputPath, excludeCovariateIds,
  deleteCovariatesSmallCount, errorOnHighCorrelation = TRUE,
  stopOnError = TRUE, fitOutcomeModel = TRUE, cdmVersion = 5)
}
\arguments{
\item{cdmDb}{The CDM database object constructed with \code{buildDatabase}}

\item{dcoList}{The list of drug-comparator-outcomes}

\item{connectionDetails}{The connectionDetails object}

\item{excludedCovariateConceptIds}{The concept Ids to exclude in Feature Extraction}

\item{cmOutputPath}{The path to the CohortMethod files}

\item{deleteCovariatesSmallCount}{The minimum cell size}

\item{errorOnHighCorrelation}{Should an error be thrown if high correlation is found in PS matching?}

\item{stopOnError}{Should the execution be interrupted if an error is thrown?}

\item{fitOutcomeModel}{Should the outcome models be fit?}

\item{cdmVersion}{The CDM version of the CDM database}
}
\details{
Executes CohortMethod for primary outcomes 
                                   by invoking \code{executeFullCm}
}
\author{
Ajit Londhe
}
